moved instrumenting classes to instrument package. removed unneeded RestTemplate infrastructure.
