+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/instrument/scheduling/TraceSchedulingAutoConfiguration.java
@@ -1,4 +1,4 @@
-package org.springframework.cloud.sleuth.scheduling;
+package org.springframework.cloud.sleuth.instrument.scheduling;
 
 /**
  * @author Spencer Gibb
