+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/instrument/web/TraceWebAspect.java
@@ -1,4 +1,4 @@
-package org.springframework.cloud.sleuth.web;
+package org.springframework.cloud.sleuth.instrument.web;
 
 import static org.springframework.cloud.sleuth.Trace.SPAN_ID_NAME;
 import static org.springframework.cloud.sleuth.Trace.TRACE_ID_NAME;
